requirements.txt
tqdm
hachoir
python-dateutil
pywin32 ; sys_platform == 'win32'